+++ Changes.md
@@ -35,6 +35,7 @@
 * LocalEnsemble: support for operating_point, operating_kind
 * Added support for Local Fusions
 * Adding organizations support for all the API calls.
+* Deprecating param byName in Local objects
 
 ### Bugs